address some of the review comments, rename files
